+++ src/main/java/com/example/domain/ecommerce/factories/ProductoFactory.java
@@ -7,5 +7,5 @@ public interface ProductoFactory {
     boolean supports(String tipo);
     Producto crearProducto(ProductDTO productDTO);
     Producto actualizar(ProductDTO productDTO, int id);
-    Producto obtener(int id);
+    Producto obtener(Long id);
 }
